+++ src/game_scene/gsgene_view.cu
@@ -239,7 +239,7 @@ void GSGeneView::onFrameRender(CL_GraphicContext* gc){
 	// draw the background
 	Physics::vec2 pos;
 	apply_transform(gc, pos);
-	m_background->draw(*gc, pos.x, pos.y);
+	m_background->draw(*gc, (f32)pos.x, (f32)pos.y);
 	
 	// draw the bullets, yes, we're cheating the numbers
 	// OOP can wait another day
@@ -247,7 +247,7 @@ void GSGeneView::onFrameRender(CL_GraphicContext* gc){
 		pos = TankBullet::get_bullet_pos(&m_bullets, i);
 		apply_transform(gc, pos);
 		if(m_bullets.state[i] != BULLET_STATE_INACTIVE){
-			m_testbullet->draw(*gc, pos.x, pos.y);
+			m_testbullet->draw(*gc, (f32)pos.x,(f32) pos.y);
 		}
 	}
 	
@@ -257,7 +257,7 @@ void GSGeneView::onFrameRender(CL_GraphicContext* gc){
 		apply_transform(gc, pos);
 		f32 rot = BasicTank::get_tank_rot(&m_tanks, m_playertank);
 		m_testtank->set_angle(CL_Angle(-rot, cl_degrees));
-		m_testtank->draw(*gc, pos.x, pos.y);
+		m_testtank->draw(*gc, (f32)pos.x, (f32)pos.y);
 	}
 
 	if(m_tanks.state[m_player2tank] != TANK_STATE_INACTIVE){
@@ -265,7 +265,7 @@ void GSGeneView::onFrameRender(CL_GraphicContext* gc){
 		apply_transform(gc, pos);
 		f32 rot = BasicTank::get_tank_rot(&m_tanks, m_player2tank);
 		m_testtank2->set_angle(CL_Angle(-rot, cl_degrees));
-		m_testtank2->draw(*gc, pos.x, pos.y);
+		m_testtank2->draw(*gc, (f32)pos.x, (f32)pos.y);
 	}
 	
 // 	if(m_tanks.state[m_player3tank] != TANK_STATE_INACTIVE){
@@ -364,7 +364,7 @@ void GSGeneView::onFrameUpdate(double dt,
 		// additional output
 		Physics::vec2 pos = m_physrunner->bodies.cur_pos.get_vec2(m_tanks.phys_id[0]);
 		f32 rot = m_physrunner->bodies.rotation[m_tanks.phys_id[0]];
-		printf("E.x: %f E.y: %f E.r: %f\n", pos.x, pos.y, rot);
+		printf("E.x: %.12f E.y: %.12f E.r: %.12f\n", pos.x, pos.y, rot);
 
 		if(m_tanks.state[m_playertank] == TANK_STATE_INACTIVE){
 			// player tank died, let's reset
@@ -425,7 +425,7 @@ void GSGeneView::onFrameUpdate(double dt,
 			// additional output
 			Physics::vec2 pos = m_physrunner->bodies.cur_pos.get_vec2(m_tanks.phys_id[0]);
 			f32 rot = m_physrunner->bodies.rotation[m_tanks.phys_id[0]];
-			printf("E.x: %f E.y: %f E.r: %f\n", pos.x, pos.y, rot);
+			printf("E.x: %.12f E.y: %.12f E.r: %.12f\n", pos.x, pos.y, rot);
 		}else if(m_tanks.state[m_playertank] == TANK_STATE_INACTIVE){
 			// player tank died, let's reset
 			cout << m_test_dist << " " << m_test_sect << " " << m_test_vect
